requirements.txt
aiojobs==0.3.0
aioredis==1.3.1
async-timeout==3.0.1
fastapi==0.66.0
fastapi-plugins==0.8.1
fastapi-utils==0.2.1
greenlet==1.1.1
h11==0.12.0
h2==4.0.0
hiredis==2.0.0
hpack==4.0.0
Hypercorn==0.11.2
hyperframe==6.0.1
numpy==1.19.5
pandas==1.3.3
Pillow==8.3.2
priority==2.0.0
psycopg2-binary==2.8.6
pydantic==1.8.2
pysftp==0.2.9
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.25
starlette==0.14.2
Sphinx==4.2.0
sphinx_rtd_theme==1.0.0
tenacity==7.0.0
toml==0.10.2
typing-extensions==3.10.0.0
wsproto==1.0.0
